+++ explore/src/main/webapp/cacheworker.js
@@ -1,0 +1,3 @@
+import { CacheIDBWorker } from '@workers/cacheworker.js';
+
+CacheIDBWorker.runWorker();
