--- explore/src/main/webapp/catalogworker.js
@@ -1,3 +1,0 @@
-import { CatalogWorker } from '@workers/catalogworker.js';
-
-CatalogWorker.runWorker();
