+++ model-tests/shared/src/test/scala/explore/model/boopickle/BoopickleSuite.scala
@@ -7,19 +7,13 @@ import eu.timepit.refined.cats._
 import eu.timepit.refined.scalacheck.string._
 import eu.timepit.refined.types.string.NonEmptyString
 import explore.boopickle.PicklerTests
-import explore.model.CatalogPicklers
 import explore.model.CatalogResults
 import explore.model.GuideStarCandidate
 import explore.model.arb.ArbCatalogResults._
-import lucuma.core.math.Angle
-import lucuma.core.math.Coordinates
-import lucuma.core.math.Declination
-import lucuma.core.math.Epoch
-import lucuma.core.math.HourAngle
-import lucuma.core.math.Parallax
-import lucuma.core.math.ProperMotion
-import lucuma.core.math.RadialVelocity
-import lucuma.core.math.RightAscension
+import explore.model.boopickle.CatalogPicklers
+import lucuma.core.model.SiderealTracking
+import lucuma.core.model.arb.ArbSiderealTracking._
+import lucuma.core.math._
 import lucuma.core.math.arb.ArbAngle._
 import lucuma.core.math.arb.ArbCoordinates._
 import lucuma.core.math.arb.ArbDeclination._
@@ -28,8 +22,6 @@ import lucuma.core.math.arb.ArbParallax._
 import lucuma.core.math.arb.ArbProperMotion._
 import lucuma.core.math.arb.ArbRadialVelocity._
 import lucuma.core.math.arb.ArbRightAscension._
-import lucuma.core.model.SiderealTracking
-import lucuma.core.model.arb.ArbSiderealTracking._
 
 class BoopickleSuite extends munit.DisciplineSuite with CatalogPicklers {
   checkAll("Pickler[Angle]", PicklerTests[Angle].pickler)
